(*) merge & rename pkg.
