+++ pkg/gomega_util/assertion.go
@@ -1,4 +1,4 @@
-package assertion
+package gomega_util
 
 import (
 	"time"
