+++ pkg/gomega_util/matcher_suite_test.go
@@ -1,4 +1,4 @@
-package matcher
+package gomega_util
 
 import (
 	"testing"
