+++ testcase/oom/tikv_test.go
@@ -15,8 +15,7 @@ import (
 	"github.com/pingcap/test-infra/sdk/resource"
 	_ "github.com/pingcap/test-infra/sdk/resource/impl/k8s"
 
-	assertion2 "github.com/pingcap/e2etest/pkg/assertion"
-	"github.com/pingcap/e2etest/pkg/matcher"
+	. "github.com/pingcap/e2etest/pkg/gomega_util"
 	. "github.com/pingcap/e2etest/pkg/util"
 	"github.com/pingcap/e2etest/pkg/workload"
 )
@@ -68,13 +67,13 @@ var _ = ParameterizedTestCase("tikvoom", func(flagSet *flag.FlagSet) {
 
 				prometheusURL, err := tc.ServiceURL(resource.Prometheus)
 				Expect(err).ShouldNot(HaveOccurred())
-				c, err := matcher.NewPrometheusClient(*prometheusURL)
+				c, err := NewPrometheusClient(*prometheusURL)
 				Expect(err).ShouldNot(HaveOccurred())
 
-				assertion2.ConsistentlyWithMetrics(matcher.PromQL(
+				ConsistentlyWithMetrics(PromQL(
 					"TiKV should not restart",
 					`rate(process_start_time_seconds{component="tikv"}[1m]) != 0`, &c),
-				).Should(matcher.PromQLEvaluatedToEmpty())
+				).Should(PromQLEvaluatedToEmpty())
 			})
 		})
 	})
